Moved semantic information to its own file.
