+++ CommonSubexpressionEliminator.cpp
@@ -24,7 +24,7 @@
 #include <functional>
 #include <boost/range/adaptor/reversed.hpp>
 #include <libevmcore/CommonSubexpressionEliminator.h>
-#include <libevmcore/Assembly.h>
+#include <libevmcore/AssemblyItem.h>
 
 using namespace std;
 using namespace dev;
@@ -246,79 +246,6 @@ ExpressionClasses::Id CommonSubexpressionEliminator::loadFromMemory(ExpressionCl
 		return m_memoryContent.at(_slot);
 	else
 		return m_memoryContent[_slot] = m_expressionClasses.find(Instruction::MLOAD, {_slot}, true, m_sequenceNumber);
-}
-
-bool SemanticInformation::breaksBasicBlock(AssemblyItem const& _item)
-{
-	switch (_item.type())
-	{
-	default:
-	case UndefinedItem:
-	case Tag:
-		return true;
-	case Push:
-	case PushString:
-	case PushTag:
-	case PushSub:
-	case PushSubSize:
-	case PushProgramSize:
-	case PushData:
-		return false;
-	case Operation:
-	{
-		if (isSwapInstruction(_item) || isDupInstruction(_item))
-			return false;
-		if (_item.instruction() == Instruction::GAS || _item.instruction() == Instruction::PC)
-			return true; // GAS and PC assume a specific order of opcodes
-		if (_item.instruction() == Instruction::MSIZE)
-			return true; // msize is modified already by memory access, avoid that for now
-		if (_item.instruction() == Instruction::SHA3)
-			return true; //@todo: we have to compare sha3's not based on their memory addresses but on the memory content.
-		InstructionInfo info = instructionInfo(_item.instruction());
-		if (_item.instruction() == Instruction::SSTORE)
-			return false;
-		if (_item.instruction() == Instruction::MSTORE)
-			return false;
-		//@todo: We do not handle the following memory instructions for now:
-		// calldatacopy, codecopy, extcodecopy, mstore8,
-		// msize (note that msize also depends on memory read access)
-
-		// the second requirement will be lifted once it is implemented
-		return info.sideEffects || info.args > 2;
-	}
-	}
-}
-
-bool SemanticInformation::isCommutativeOperation(AssemblyItem const& _item)
-{
-	if (_item.type() != Operation)
-		return false;
-	switch (_item.instruction())
-	{
-	case Instruction::ADD:
-	case Instruction::MUL:
-	case Instruction::EQ:
-	case Instruction::AND:
-	case Instruction::OR:
-	case Instruction::XOR:
-		return true;
-	default:
-		return false;
-	}
-}
-
-bool SemanticInformation::isDupInstruction(AssemblyItem const& _item)
-{
-	if (_item.type() != Operation)
-		return false;
-	return Instruction::DUP1 <= _item.instruction() && _item.instruction() <= Instruction::DUP16;
-}
-
-bool SemanticInformation::isSwapInstruction(AssemblyItem const& _item)
-{
-	if (_item.type() != Operation)
-		return false;
-	return Instruction::SWAP1 <= _item.instruction() && _item.instruction() <= Instruction::SWAP16;
 }
 
 CSECodeGenerator::CSECodeGenerator(
